Multiple changes: A simpler model containing two populations (a Poisson population as input and an output population of Izhikevich phasic bursting type) is included as an example. README modified to include this function call and corresponding Makefile modification.
NNmodel now includes vector<int> receivesInputCurrent flag which defines whether the model receives an external current injection or not. If this is the case, calcNeurons is generated with an additional arg "float *inputI"model.neuronName[i]"[localID]" , which contains the input current to be added to ISyn (needs to be tested).

Redefinition of DT is removed from modelSpec.h.

Izhikevich model can now be called correctly with any time step. However, a warning is thrown if DT is not equal to 1ms.