Applying Thomas' last bug fix in the v2.0_benchmark branch.
I have identified and fixed two very tricky bugs in the synapse kernel.
1.
The mechanism with numSpikeSubsets was flawed. First of all, the solution that was
calculating numSpikeSubsets with float converted integer variables and then converted back
to unsigned integer was odd and might have given the wrong result for lscnt == BLOCKSZ_SYN
depending on the behaviour of fceil and the rounding situation.
Instead I have now used the formula
(lscnt+BLOCKSZ_SYN-1)/ BLOCKSZ_SYN
which gives 1 for lscnt from 1 ... BLOCKSZ_SYN as intended and so on.
It's a bit dicey for lscnt == 0, where it evaluates to -1/BLOCKSZ_SYN which is 0, but it
goes through a signed intermediate value. I am a bit uneasy about this but it seems to
work correctly.
The other problem under this heading was that subsequently the formula for lmax was not
correct, giving lmax == 0 in the (unlikely but possible) event of lscnt == BLOCKSZ_SYN.
I have now amended with lmax = ((lscnt-1) % BLOCKSZ_SYN) +1; which should always have the
correct behaviour.

2.
The bug that was causing the different number of DN output spikes in the
MBody_userdef_benchmark_project was very tricky. I believe it was due to incomplete
synchronization of threads in the synapse kernel code. We had a section of code where
first (lid < num_post) was tested and inside the if clause a loop was executed that
contained a __syncthreads() call. This wouldn't have synchronised the other threads not
within the if clause and for some reason that I have not completely thought through
this has caused problems where contributions to linSyn and eventually dd_inSynXXYY were
missed; This was especially true for block sizes where the shLg mechanism was employed
rather than the atomicAdd version but I do believe (this is subject to confirmation) that
even the atomicAdd version was missing contributions to dd_inSynXXYY.
I have now pulled the if clause into the loop such that all threads do the loop and the
__syncthreads() in it. As far as I can tell from limited testing this seems to solve the
issue of missing output spikes and the number of output spikes in the
MBody_userdef_benchmark_project example appears now to be fully reproducible between runs
and for any valid block size.

We should consider merging this bug fix into development soon & merging development into
master to create a service release.

Conflicts:
	userproject/MBody_userdef_benchmark_project/model/classol_sim.cu
	userproject/MBody_userdef_benchmark_project/model/map_classol.cc