This is mainly a bug fix for incorrect scoping of local variables in kernels when support code was used. In brief, the support code would live in separate namespaces and to use it, we introduced local variable scopes that contained "using namespace ***". These led to the definition of "oldSpike" to be in its own scope and it would be undefined later. In the new version, local scopes are reduced to a minimum while still keeping the functionality of "using namspace" for support code. The existing tests and examples run but there are not tests with support code, so this doesn't mean much.
On the side, there was a problem discovered for Mac OS X El Capitan. The new SIP (System Integrity Protection) introduced in Mac OS X El Capitan means that DYLD_LIBRARY_PATH is sqashed in sub-shells. We depended on this variable to correctly link to libcudart.dylib. This was because libcudart.*.dylib is compiled with install location/name @rpath/libcuart.*.dylib. The best fix of the combination of these two problems is to define rpath during linking stage of our meta-compiler and our user executables to be the correct path to
libcudart.*.dylib.
This has now been done in the appropriate makefiles.