There was a problem with the new binomialInverseCDF in that many of the terms added to make up the CDF are flushed to 0 even for long double floating point precision. For more challenging examples, this means that the CDF is under-estimated and in the worst cases, the desired CDF is never reached and the function returned the inverse "n". Here I have introduced a tweak (hack?) that makes use of the fact that the missing contributions are for small k << n*p and for large k >> n*p. If the desired CDF is small, we use the usual algorithm and the missing values for small k aren't that important as the contributions of each term to the sum increase in size rapidly and the last contribution easily outweighs all missing conributions, such that the error made is only on the order of 1.
To get the same benefits of rising contributions for the case of large
desired CDFs (desired CDF > p), we calculate the sum of probabilities for k
starting at n and descending to 0 and call the inverse CDF result, when this
sum has surpassed (1- desired CDF). With this tweak, we get the same benefits
and the error incurred due to missing terms for large k causes only minor
errors on the inverse CDF value.

I have tested against SciPy's PPF function for n=1e6, p=0.02 and desired CDFS:
1e-10 to 1e-20, and (1-1e-10) ... (1-1e-16) and the results differed at most
by 1.